add interfaces list to generatedTypes
